Project structure: move classes into packages, add package and import statements for lazysourcea, task, storage, ui, and parser
