--- src/main/java/EmptyTaskListException.java
@@ -1,5 +1,0 @@
-public class EmptyTaskListException extends Exception {
-    public EmptyTaskListException(String message) {
-        super(message);
-    }
-}
