+++ src/main/java/lazysourcea/EmptyTaskListException.java
@@ -1,0 +1,7 @@
+package lazysourcea;
+
+public class EmptyTaskListException extends Exception {
+    public EmptyTaskListException(String message) {
+        super(message);
+    }
+}
