+++ src/main/java/lazysourcea/parser/Parser.java
@@ -1,3 +1,5 @@
+package lazysourcea.parser;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
